Revert "Add overall structure (#5)" (#6)
This reverts commit 13bb13671ab25dde0bbdd7ff1d3313ce6cbadb81.
