+++ gccc-site/package.json
@@ -8,21 +8,20 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
+    "register-service-worker": "^1.5.2",
+    "vue": "^2.6.6",
+    "vue-class-component": "^6.0.0",
+    "vue-property-decorator": "^7.0.0",
     "@vue/cli-plugin-babel": "^3.4.0",
     "@vue/cli-plugin-pwa": "^3.4.0",
     "@vue/cli-plugin-typescript": "^3.4.0",
     "@vue/cli-service": "^3.4.0",
-    "node-sass": "^4.11.0",
-    "register-service-worker": "^1.5.2",
-    "sass-loader": "^7.1.0",
     "typescript": "^3.0.0",
-    "vue": "^2.6.6",
-    "vue-class-component": "^6.0.0",
-    "vue-property-decorator": "^7.0.0",
-    "vue-router": "^3.0.1",
     "vue-template-compiler": "^2.5.21"
   },
-  "devDependencies": {},
+  "devDependencies": {
+
+  },
   "postcss": {
     "plugins": {
       "autoprefixer": {}
