+++ gccc-site/src/App.vue
@@ -1,21 +1,23 @@
 <template>
   <div id="app">
     <Header/>
-    <div class="header-padding"/>
-    <router-view/>
+    <MainContent/>
     <Footer/>
   </div>
 </template>
 
 <script lang="ts">
-import { Component, Prop, Vue } from "vue-property-decorator";
-import Header from "@/components/Header.vue";
-import Footer from "@/components/Footer.vue";
+import { Component, Vue } from "vue-property-decorator";
+
+import Header from "./components/Header.vue";
+import Footer from "./components/Footer.vue";
+import MainContent from "./components/MainContent.vue";
 
 @Component({
   components: {
     Header,
     Footer,
+    MainContent,
   },
 })
 export default class App extends Vue {}
@@ -36,9 +38,5 @@ export default class App extends Vue {}
 body {
   padding: 0;
   margin: 0;
-}
-
-.header-padding {
-  height: 6vh;
 }
 </style>
