--- gccc-site/src/components/contents/NewsContent.vue
@@ -1,29 +1,0 @@
-<template>
-	<section id="news-content">
-    <p>New page is not implemented yet :/</p>
-	</section>
-</template>
-
-<script lang="ts">
-// import decorator below
-import { Component, Prop, Vue } from "vue-property-decorator";
-
-// import components below
-
-@Component({
-  components: {
-    // use components here
-  },
-})
-
-export default class NewsContent extends Vue {
-  public created() {
-    const t: string = "";
-  }
-}
-</script>
-
-
-<style lang="scss" scoped>
-// scoped is affect only this components
-</style>
