--- gccc-site/src/components/contents/WorksContent.vue
@@ -1,29 +1,0 @@
-<template>
-	<section id="works-content">
-    <p>Works page is not implemented yet :/</p>
-	</section>
-</template>
-
-<script lang="ts">
-// import decorator below
-import { Component, Prop, Vue } from "vue-property-decorator";
-
-// import components below
-
-@Component({
-  components: {
-    // use components here
-  },
-})
-
-export default class WorksContent extends Vue {
-  public created() {
-    const t: string = "";
-  }
-}
-</script>
-
-
-<style lang="scss" scoped>
-// scoped is affect only this components
-</style>
