+++ gccc-site/src/components/maincontents/About.vue
@@ -19,6 +19,7 @@ export default class About extends Vue {}
 
 <style scope>
 #about {
+  padding-top: 5vh;
   border-bottom: solid thin gray;
 }
 </style>
