--- gccc-site/src/views/Home.vue
@@ -1,17 +1,0 @@
-<template>
-  <div class="home">
-    <HomeContent/>
-  </div>
-</template>
-
-<script>
-// @ is an alias to /src
-import HomeContent from '@/components/contents/HomeContent.vue'
-
-export default {
-  name: 'home',
-  components: {
-    HomeContent,
-  },
-}
-</script>
