--- gccc-site/src/views/News.vue
@@ -1,17 +1,0 @@
-<template>
-  <div class="News">
-    <NewsContent/>
-  </div>
-</template>
-
-<script>
-// @ is an alias to /src
-import NewsContent from '@/components/contents/NewsContent.vue'
-
-export default {
-  name: 'news',
-  components: {
-    NewsContent,
-  },
-}
-</script>
